Conflicts:
	src/Makefile.am
	src/backend/bridge/dml/mapper/mapper_utils.cpp
	src/backend/executor/delete_executor.cpp
	src/backend/executor/insert_executor.cpp
	src/backend/executor/update_executor.cpp
	src/backend/storage/tile_group.cpp
	src/backend/storage/tile_group.h
	src/backend/storage/tile_group_header.h
	src/postgres/backend/postmaster/peloton.cpp
	src/postgres/backend/tcop/postgres.cpp

Former-commit-id: f41f140267d54f4aa5eefcf98548a2b6ac5dceeb

+++ src/backend/common/types.cpp
@@ -107,7 +107,7 @@ std::string ValueTypeToString(ValueType type) {
     case VALUE_TYPE_BIGINT:
       return "BIGINT";
     case VALUE_TYPE_DOUBLE:
-      return "FLOAT";
+      return "DOUBLE";
     case VALUE_TYPE_VARCHAR:
       return "VARCHAR";
     case VALUE_TYPE_VARBINARY:
@@ -134,7 +134,7 @@ ValueType StringToValueType(std::string str) {
     return VALUE_TYPE_INTEGER;
   } else if (str == "BIGINT") {
     return VALUE_TYPE_BIGINT;
-  } else if (str == "FLOAT") {
+  } else if (str == "DOUBLE") {
     return VALUE_TYPE_DOUBLE;
   } else if (str == "STRING") {
     return VALUE_TYPE_VARCHAR;
